Fix firefly spawn density config under 100 not spawning any illuminations

+++ src/main/java/ladysnake/illuminations/client/data/IlluminationData.java
@@ -39,7 +39,7 @@ public class IlluminationData {
     }
 
     public boolean shouldAddParticle(Random random) {
-        int density = Config.getDensity()/100;
+        float density = Config.getDensity()/100f;
         return random.nextFloat() <= this.chance * density;
     }
 }
